Reduce number of inner constructors
Best practice in Julia is to have as few inner constructors as possible, which
are primarily there to enforce invariants (in our case: to update ring
refcounts and set finalizers). The other constructors then can be implemented
as outer constructors which delegate to the remaining inner ones. This reduces
code duplication.